requirements/base.txt
python-magic-bin==0.4.14;platform_system=="Windows"
argon2-cffi==23.1.0
bleach==6.1.0
crispy-bootstrap5==2024.2
django==4.2.15
django-allauth==64.2.0
django-allauth[mfa]==64.2.0
django-anymail[mailgun]==11.1
django-braces==1.15.0
django-cjswidget==0.4.1
django-cleanup==8.1.0
django-crispy-forms==2.3
django-csp==3.8
django-csp-helpers>=0.7.1
django-environ==0.11.2
django-mail-templated-simple==3.2
django-markdownfield==0.11.0
django-oidc-provider==0.8.2
django-otp==1.5.2
django-published>=0.8.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-rq==2.10.2
django-slugger==1.1.2
django-storages[azure]==1.14.4
django-structured-data==0.6.0
django-taggit==6.0.0
django-versatileimagefield==3.1
django-webpack-loader==3.1.1
fido2==1.1.3
foxtail-blog>=0.24.2
foxtail-contact>=0.6.0
whitenoise==6.7.0
gitpython==3.1.43
pytz==2024.1
gunicorn==23.0.0
markdown==3.7
psycopg2==2.9.10;platform_system!="Windows"
psycopg2-binary==2.9.10;platform_system=="Windows"
pymdown-extensions==10.9
requests==2.32.3
requests-oauthlib==2.0.0
PyJWT==2.9.0
rules==3.5
rq==1.16.2
sentry-sdk==2.13.0
the_big_username_blacklist==1.5.4
